requirements.txt
pandas>=2.0
openpyxl>=3.1
python-dateutil>=2.8
python-dotenv>=1.0
